+++ projects/ngx-eutrepe-scroll-to/src/lib/windowToken/window-token.ts
@@ -1,6 +1,6 @@
 import { InjectionToken } from '@angular/core';
 
-export const WINDOW = new InjectionToken<Window>('WINDOW', {
+export const WINDOW_SCROLL_TO = new InjectionToken<Window>('WINDOW_SCROLL_TO', {
   providedIn: 'root',
   factory: () => window
 });