+++ src/app/app.component.ts
@@ -1,5 +1,5 @@
 import { Component, AfterViewInit, ViewChild, ElementRef } from '@angular/core';
-import { NgxEutrepeScrollToService } from '@eutrepe/scroll-to';
+import { NgxEutrepeScrollToService, IScrollToConfig } from 'ngx-eutrepe-scroll-to';
 
 @Component({
   selector: 'eutrepe-root',
@@ -12,22 +12,37 @@ export class AppComponent implements AfterViewInit {
 
   constructor(private ngxEutrepeScrollToService: NgxEutrepeScrollToService){}
 
+  private config: IScrollToConfig = {
+    easing: 'linear',
+    duration: 3000,
+    offset: 100,
+    onStart: this.onStart,
+    onEnd: this.onEnd,
+    onStartParams: ['some_text', 1, true],
+    onEndParams: [100, false],
+    onBreakParams: ['Eutrepe'],
+    onBreak: this.onBreak
+    }
+
   ngAfterViewInit() {
-    this.ngxEutrepeScrollToService.scrollTo(this.test.nativeElement, {
-      easing: 'linear',
-      duration: 3000,
-      offset: 100,
-      onStart: this.onStart,
-      onEnd: this.onEnd
-    });
+    this.ngxEutrepeScrollToService.scrollTo(this.test.nativeElement, this.config);
   }
 
 
-  onStart() {
+  onStart(arg1, arg2, arg3) {
+    console.log(arg1); // 'some_text'
+    console.log(arg2); // 1
+    console.log(arg3); // true
     console.log('start scrolling');
   }
 
-  onEnd() {
+  onEnd(arg1, arg2) {
+    console.log(arg1); // 100
+    console.log(arg2); // false
     console.log('finish scrolling');
+  }
+
+  onBreak(arg1) {
+    console.log(arg1);  // Eutrepe
   }
 }