+++ src/app/app.module.ts
@@ -1,10 +1,10 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 
-import { NgxEutrepeScrollToDirective } from '@eutrepe/scroll-to';
+import { NgxEutrepeScrollToDirective } from 'ngx-eutrepe-scroll-to';
 
 import { AppComponent } from './app.component';
-import { WINDOW } from '@eutrepe/scroll-to';
+import { WINDOW_SCROLL_TO } from 'ngx-eutrepe-scroll-to';
 
 @NgModule({
   declarations: [
@@ -15,7 +15,7 @@ import { WINDOW } from '@eutrepe/scroll-to';
     BrowserModule
   ],
   providers: [
-    {provide: WINDOW, useValue: window}
+    {provide: WINDOW_SCROLL_TO, useValue: window}
   ],
   bootstrap: [AppComponent]
 })